visa: deprecate the visa module
Address https://github.com/pyvisa/pyvisa/issues/447
