+++ visa.py
@@ -13,6 +13,15 @@
 
 from __future__ import division, unicode_literals, print_function, absolute_import
 
+import warnings
+warnings.warn(("The visa module provided by PyVISA is being deprecated. "
+               "You can replace `import visa` by `import pyvisa as visa` "
+               "to achieve the same effect.\n\n"
+               "The reason for the deprecation is the possible conflict with "
+               "the visa package provided by the "
+               "https://github.com/visa-sdk/visa-python which can result in "
+               "hard to debug situations."), FutureWarning)
+
 from pyvisa import logger, __version__, log_to_screen, constants
 from pyvisa.highlevel import ResourceManager
 from pyvisa.errors import (Error, VisaIOError, VisaIOWarning, VisaTypeError,
@@ -20,36 +29,7 @@ from pyvisa.errors import (Error, VisaIOError, VisaIOWarning, VisaTypeError,
                            InvalidSession, LibraryError)
 # This is needed to registry all resources.
 from pyvisa.resources import Resource
-
-def visa_main(command=None):
-    import argparse
-    parser = argparse.ArgumentParser(description='PyVISA command-line utilities')
-
-    parser.add_argument('--backend', '-b', dest='backend', action='store', default=None,
-                        help='backend to be used (default: ivi)')
-
-    if not command:
-        subparsers = parser.add_subparsers(title='command', dest='command')
-
-        info_parser = subparsers.add_parser('info', help='print information to diagnose PyVISA')
-
-        console_parser = subparsers.add_parser('shell', help='start the PyVISA console')
-
-    args = parser.parse_args()
-    if command:
-        args.command = command
-    if args.command == 'info':
-        from pyvisa import util
-        util.get_debug_info()
-    elif args.command == 'shell':
-        from pyvisa import shell
-        shell.main('@' + args.backend if args.backend else '')
-
-def visa_shell():
-    visa_main('shell')
-
-def visa_info():
-    visa_main('info')
+from pyvisa.cmd_line_tools import visa_main
 
 if __name__ == '__main__':
     visa_main()
